# apps/api/src/openapi.yaml
openapi: 3.1.0
info:
  title: ashim API
  version: 1.15.9
  description: |
    REST API for ashim, a self-hosted image processing platform with 30+ tools.

    ## Authentication

    Most endpoints require authentication via one of two methods:

    1. **Session cookie** — Call `POST /api/auth/login` with username and password. The response includes a `token` field. Pass it as `Authorization: Bearer <token>` on subsequent requests.

    2. **API key** — Generate a key via the Settings UI or `POST /api/v1/api-keys`. Keys are prefixed with `si_`. Pass as `Authorization: Bearer si_...`.

    Endpoints marked with a lock icon require authentication. Admin-only endpoints are noted in their description.
  license:
    name: AGPL-3.0
    url: https://github.com/ashim-hq/ashim/blob/main/LICENSE
servers:
  - url: /
    description: Current instance
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Session token from login or API key (prefixed with si_)
  schemas:
    Error:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: Invalid image format
...
    UnauthorizedError:
      type: object
      properties:
        statusCode:
          type: integer
          example: 401
        error:
          type: string
          example: Unauthorized
        message:
          type: string
          example: Authentication required
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/files:
    get:
      tags: [Files]
      summary: List saved files
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of saved files
          content:
            application/json:
              schema:
                type: object
                properties:
                  files:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        filename:
                          type: string
                        size:
                          type: integer
                        contentType:
                          type: string
                        createdAt:
                          type: string
                          format: date-time
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
    delete:
      tags: [Files]
      summary: Bulk delete saved files
      description: Delete files and their entire version chains. Non-admin users can only delete their own files.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids]
              properties:
                ids:
                  type: array
                  items:
                    type: string
                  description: File IDs to delete (deletes entire version chain for each)
      responses:
        "200":
          description: Deletion result
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: integer
                    description: Total number of records deleted across all version chains
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
  /api/v1/files/upload:
    post:
      tags: [Files]
      summary: Save file to library
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Saved file
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  filename:
                    type: string
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
  /api/v1/files/save-result:
    post:
      tags: [Files]
      summary: Save a processing result to library
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [jobId, filename]
              properties:
                jobId:
                  type: string
                filename:
                  type: string
      responses:
        "200":
          description: Saved file
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  filename:
                    type: string
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
  /api/v1/files/{id}:
    get:
      tags: [Files]
      summary: Get file metadata
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Saved file UUID
          schema:
            type: string
      responses:
        "200":
          description: File metadata
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  filename:
                    type: string
                  size:
                    type: integer
                  contentType:
                    type: string
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
  /api/v1/files/{id}/download:
    get:
      tags: [Files]
      summary: Download a saved file
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Saved file UUID
          schema:
            type: string
      responses:
        "200":
          description: Image binary
          content:
            image/*:
              schema:
                type: string
                format: binary
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
  /api/v1/files/{id}/thumbnail:
    get:
      tags: [Files]
      summary: Get file thumbnail
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Saved file UUID
          schema:
            type: string
      responses:
        "200":
          description: Image binary
          content:
            image/*:
              schema:
                type: string
                format: binary
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"

  # ─── Auth ─────────────────────────────────────────────────────────────────
...